按钮控件，单选控件，修复绑定问题

+++ samples/webgpu/src/appbase.h
@@ -279,7 +279,7 @@ private:
 	{
 		auto ve = cWnd->GetWindowUI()->_children[0];
 		onReloadedEvent(ve);
-		ve->_pseudoMask |= (int)PseudoStates::Root;
+		ve->_pseudoMask |= PseudoStates::Root;
 	}
 };
